requirement.txt
```
requests
beautifulsoup4
tqdm
pandas
numpy
faiss-cpu
sentence-transformers
fastapi
uvicorn[standard]
google-generativeai
pydantic
streamlit
rank_bm25
torch
transformers
playwright
openpyxl
groq
python-dotenv
```